requirements.txt
click==8.1.8
dataclasses-json==0.5.7
markdown-it-py==3.0.0
marshmallow==3.23.1
marshmallow-enum==1.5.1
mdurl==0.1.2
mypy-extensions==1.0.0
packaging==24.2
psutil==6.1.0
Pygments==2.18.0
pyre-extensions==0.0.32
python-frontmatter==1.1.0
PyYAML==5.1
pytest==8.3.5
rich==13.9.4
tabulate==0.9.0
TestSlide==2.7.1
typeguard==2.13.3
typing-inspect==0.9.0
typing_extensions==4.12.2
